+++ fugitive-darkness-common/src/main/java/com/zer0s2m/fugitivedarkness/common/dto/ContainerGitRepoSearchFilters.java
@@ -10,7 +10,9 @@ public record ContainerGitRepoSearchFilters(
 
         List<String> excludeExtensionFiles,
 
-        String patternForIncludeFile
+        String patternForIncludeFile,
+
+        String patternForExcludeFile
 
 ) {
 }