+++ fugitive-darkness-provider/src/main/java/com/zer0s2m/fugitivedarkness/provider/impl/GitRepoImpl.java
@@ -163,6 +163,10 @@ public class GitRepoImpl implements GitRepo {
                             commandGrep.setPatternForIncludeFile(filterSearch.getPatternForIncludeFile());
                         }
 
+                        if (filterSearch.getPatternForExcludeFile() != null) {
+                            commandGrep.setPatternForExcludeFile(filterSearch.getPatternForExcludeFile());
+                        }
+
                         final List<ContainerInfoSearchFileGitRepo> searchResult = commandGrep.callGrep();
 
                         searchFileGitRepos.add(new ContainerInfoSearchGitRepo(